Implement basic compilation pipeline
This is the first commit in this repository. This initial version
contains very basic functionality for defining functions and doing basic
operations with basic types. Here is the current feature set:

- Defining functions with parameters
- Primitive types: i32, f32 and bool
- Binary `+` for i32, binary `*` for f32 and unary `!` for bool

Compilation failure results in a mildly helpful error message.
Successful compilation results in producing textual LLVM code and
sending it to `clang`.