Simplify Error types by forgoing Display impl
Display is nice but compiler::Error requires the text string, which
can't be passed into fmt, meaning we had to use a field. Also if that
field were a &str, we would get lifetime complications. With this
change, we can easily just pass a &str. We can also remove a panic,
because we can ensure that compiler::Error has text to read.

I also changed crate::Error in the same way just to make things
consistent.