Invert Namespace type
Basically, instead of:

    namespace: HashMap<String, Vec<(Symbol, Namespace)>>

We now have:

    symbol: Symbol
    namespace: HashMap<String, Vec<Namespace>>

Which is really just the same thing, but you remove a tuple. It also
ensures a symbol is present and says that the symbol "comes first", as a
kind of header. This means the global namespace which previously had no
symbol, now has a Symbol::Module, which is currently just a placeholder.