Add support for all primitive types and operators
Technically not all, because usize and isize are missing.

This commit basically just abstracts over existing mechanisms to support
more types and operators. We are using a lot of (hacky?) macros to
prevent huge amounts of repetition.

The entire idea behind this design is that operators are really just
overloaded functions and should be treated accordingly. That's why we
added support for function overloading first.

The testing here is very thin. There's too much to test.