Give proper types to literals
Another problem we can solve with overloading! In the same way that
operators are actually just overloaded functions, literals are just
overloaded values. An int literal is just a regular symbol that happens
to correspond to multiple types (i8, i16, ...), and so all we have to do
is choose the type that works.