Rewrite test
What should this code do?

    (f (i32) 0)
    (f (i32) (f))
    (main (i32) (f))

I think the function on the second line should be recursive, rather than
calling the other function. f1 (the second f) is in scope in the body of
f1, so the normal overloading rules apply, i.e. when there are multiple
valid functions, the most recent one is used.